docker/user-service/requirements.txt
pymongo
opentelemetry-api
opentelemetry-sdk
